SceneWriter, Loader abstraction. Resolved some component schema issues.

+++ SynapseEngine/Engine/Component/Physics/RigidBodyComponent.h
@@ -11,7 +11,7 @@ namespace Syn
         float mass = 1.0f;
         float friction = 0.2f;
         float restitution = 0.0f;
-        uint16_t layer = 1;
+        uint32_t layer = 1;
 
         PhysicsBodyID bodyID = INVALID_BODY_ID;
     };
